Making the get/setValue in Event an interface
Template get/set call a virtual method which uses std::any to hide
types.
This creates an interface and reduces the need for casting. Casting was my main issue witht he previouse method. A consumer should not need to know the underlying event typr.
Type errors are cought at run time. Not the nicests but the best I have so far.